+++ paddle/ir/builtin_dialect.cc
@@ -14,7 +14,6 @@
 
 #include "paddle/ir/builtin_dialect.h"
 #include "paddle/ir/builtin_attribute.h"
-#include "paddle/ir/builtin_op.h"
 #include "paddle/ir/builtin_type.h"
 
 namespace ir {
@@ -27,7 +26,6 @@ void BuiltinDialect::initialize() {
   // Register all built-in types defined in builtin_type.h.
   RegisterTypes<GET_BUILT_IN_TYPE_LIST>();
   RegisterAttributes<GET_BUILT_IN_ATTRIBUTE_LIST>();
-  // RegisterOperations<GET_BUILT_IN_OP_LIST>();
 }
 
 }  // namespace ir
